+++ Source/Bake/ValueObjects/Recipes/Go/GoBuildRecipe.cs
@@ -36,10 +36,7 @@ namespace Bake.ValueObjects.Recipes.Go
         public string WorkingDirectory { get; [Obsolete] set; }
 
         [YamlMember]
-        public GoOs Os { get; [Obsolete] set; }
-
-        [YamlMember]
-        public GoArch Arch { get; [Obsolete] set; }
+        public Platform Platform { get; [Obsolete] set; }
 
         [Obsolete]
         public GoBuildRecipe() { }
@@ -47,16 +44,14 @@ namespace Bake.ValueObjects.Recipes.Go
         public GoBuildRecipe(
             string output,
             string workingDirectory,
-            GoOs os,
-            GoArch arch,
+            Platform platform,
             params Artifact[] artifacts)
             : base(artifacts)
         {
 #pragma warning disable CS0612 // Type or member is obsolete
             Output = output;
             WorkingDirectory = workingDirectory;
-            Os = os;
-            Arch = arch;
+            Platform = platform;
 #pragma warning restore CS0612 // Type or member is obsolete
         }
     }